Logging client fd before closing it.
This proves that the correct fd is being closed and I think I now
understand what's happening. I was previously focusing too much on the
switch that handles each states that I forgot to look at the whole loop.
Each iteration grabs the client pointer associated with the latest IO
"completion" (if I'm using the term correctly). When the server
successfully echoes what the client sent, i.e., when it's time to handle
the "send" state, the correct client pointer associated with the completion
event is "grabbed" in the loop before the state handling part. That is
how the correct fd is obtained and closed, along with the freeing of the
memory "associated" to the corresponding pointer.